chore: move Roles to EE
